Move change_pin-function to its own trait
