+++ libwebauthn/src/pin.rs
@@ -1,8 +1,11 @@
+use std::time::Duration;
+
 use super::transport::error::Error;
 
 use aes::cipher::{block_padding::NoPadding, BlockDecryptMut};
 use async_trait::async_trait;
 use cbc::cipher::{BlockEncryptMut, KeyIvInit};
+use ctap_types::cose;
 use hkdf::Hkdf;
 use hmac::Mac;
 use p256::{
@@ -13,9 +16,15 @@ use rand::{rngs::OsRng, thread_rng, Rng};
 use sha2::{Digest, Sha256};
 use tracing::{error, info, instrument, warn};
 use x509_parser::nom::AsBytes;
-use ctap_types::cose;
 
-use crate::proto::{ctap2::Ctap2PinUvAuthProtocol, CtapError};
+use crate::{
+    proto::{
+        ctap2::{Ctap2, Ctap2ClientPinRequest, Ctap2PinUvAuthProtocol},
+        CtapError,
+    },
+    transport::Channel,
+    webauthn::{obtain_pin, obtain_shared_secret, select_uv_proto},
+};
 
 type Aes256CbcEncryptor = cbc::Encryptor<aes::Aes256>;
 type Aes256CbcDecryptor = cbc::Decryptor<aes::Aes256>;
@@ -176,8 +185,8 @@ impl ECPrivateKeyPinUvAuthProtocol for PinUvAuthProtocolOne {
 }
 
 impl<P> ECDHPinUvAuthProtocol for P
-    where
-        P: ECPrivateKeyPinUvAuthProtocol,
+where
+    P: ECPrivateKeyPinUvAuthProtocol,
 {
     #[instrument(skip_all)]
     fn encapsulate(
@@ -418,4 +427,105 @@ pub fn hkdf_sha256(salt: &[u8], ikm: &[u8], info: &[u8]) -> Vec<u8> {
     hk.expand(info, &mut okm)
         .expect("32 is a valid length for Sha256 to output");
     Vec::from(okm)
+}
+
+#[async_trait]
+pub trait PinManagement {
+    async fn change_pin(
+        &mut self,
+        pin_provider: &Box<dyn PinProvider>,
+        new_pin: String,
+        timeout: Duration,
+    ) -> Result<(), Error>;
+}
+
+#[async_trait]
+impl<C> PinManagement for C
+where
+    C: Channel,
+{
+    async fn change_pin(
+        &mut self,
+        pin_provider: &Box<dyn PinProvider>,
+        new_pin: String,
+        timeout: Duration,
+    ) -> Result<(), Error> {
+        let get_info_response = self.ctap2_get_info().await?;
+
+        // If the minPINLength member of the authenticatorGetInfo response is absent, then let platformMinPINLengthInCodePoints be 4.
+        if new_pin.as_bytes().len() < get_info_response.min_pin_length.unwrap_or(4) as usize {
+            // If platformCollectedPinLengthInCodePoints is less than platformMinPINLengthInCodePoints then the platform SHOULD display a "PIN too short" error message to the user.
+            // TODO: New error for "PIN too short" vs. "PIN too long"?
+            return Err(Error::Ctap(CtapError::PINPolicyViolation));
+        }
+
+        // If the byte length of "newPin" is greater than the max UTF-8 representation limit of 63 bytes, then the platform SHOULD display a "PIN too long" error message to the user.
+        if new_pin.as_bytes().len() >= 64 {
+            // TODO: New error for "PIN too short" vs. "PIN too long"?
+            return Err(Error::Ctap(CtapError::PINPolicyViolation));
+        }
+
+        let current_pin = match get_info_response.options.as_ref().unwrap().get("clientPin") {
+            // Obtaining the current PIN, if one is set
+            Some(true) => Some(obtain_pin(self, pin_provider, timeout).await?),
+
+            // No PIN set yet
+            Some(false) => None,
+
+            // Device does not support PIN
+            None => {
+                // TODO: What's the correct error here? Spec is unclear. Maybe just not early-return and let the device error out for us.
+                return Err(Error::Ctap(CtapError::UnsupportedOption));
+            }
+        };
+
+        // In preparation for obtaining pinUvAuthToken, the platform:
+        // * Obtains a shared secret.
+        let uv_proto = select_uv_proto(&get_info_response).await?;
+        let (public_key, shared_secret) = obtain_shared_secret(self, &uv_proto, timeout).await?;
+
+        // paddedPin is newPin padded on the right with 0x00 bytes to make it 64 bytes long. (Since the maximum length of newPin is 63 bytes, there is always at least one byte of padding.)
+        let mut padded_new_pin = new_pin.as_bytes().to_vec();
+        padded_new_pin.resize(64, 0x00);
+
+        // newPinEnc: the result of calling encrypt(shared secret, paddedPin) where
+        let new_pin_enc = uv_proto.encrypt(&shared_secret, &padded_new_pin)?;
+
+        let req = match current_pin {
+            Some(curr_pin) => {
+                // pinHashEnc: The result of calling encrypt(shared secret, LEFT(SHA-256(curPin), 16)).
+                let pin_hash = pin_hash(&curr_pin);
+                let pin_hash_enc = uv_proto.encrypt(&shared_secret, &pin_hash)?;
+
+                // pinUvAuthParam: the result of calling authenticate(shared secret, newPinEnc || pinHashEnc)
+                let uv_auth_param = uv_proto.authenticate(
+                    &shared_secret,
+                    &[new_pin_enc.as_slice(), pin_hash_enc.as_slice()].concat(),
+                );
+
+                Ctap2ClientPinRequest::new_change_pin(
+                    uv_proto.version(),
+                    &new_pin_enc,
+                    &pin_hash_enc,
+                    public_key,
+                    &uv_auth_param,
+                )
+            }
+            None => {
+                // pinUvAuthParam: the result of calling authenticate(shared secret, newPinEnc).
+                let uv_auth_param = uv_proto.authenticate(&shared_secret, &new_pin_enc);
+
+                Ctap2ClientPinRequest::new_set_pin(
+                    uv_proto.version(),
+                    &new_pin_enc,
+                    public_key,
+                    &uv_auth_param,
+                )
+            }
+        };
+
+        // On success, this is an all-empty Ctap2ClientPinResponse
+        let _ = self.ctap2_client_pin(&req, timeout).await?;
+        Ok(())
+    }
 }
